move to current MTIs
git-svn-id: svn+ssh://svn.code.sf.net/p/openlcb/svn/trunk/prototypes/java@529 87536cb5-86f1-4cdc-8fc6-3f0f71bcd267

+++ src/org/openlcb/can/CanFrame.java
@@ -16,6 +16,6 @@ package org.openlcb.can;
 
 public interface CanFrame {
 
-    public long getHeader();
+    public int getHeader();
     
 }
